requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
tqdm==4.67.0
urllib3==2.2.3
yarl==1.17.1
colorama~=0.4.6